move jsonrequest to interactive searcher
