+++ lab11/app/src/main/java/com/example/tjelvarguo/lab1/lab3/Lab3.java
@@ -26,72 +26,9 @@ import java.util.ArrayList;
 
 public class Lab3 extends Activity {
 
-    private PopUpList popUpList;
-    private EditText searchBar;
-    private InteractiveSearcher interactiveSearcher;
-    private int requestId;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.lab3);
-        interactiveSearcher = findViewById(R.id.interactiveSearcher);
-        popUpList = interactiveSearcher.getPopUpList();
-        searchBar = interactiveSearcher.getSearchBar();
-
-        final ArrayList<String> names = new ArrayList<>();
-
-        final RequestQueue queue = Volley.newRequestQueue(this);
-
-
-
-        searchBar.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
-            }
-
-            @Override
-            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-                popUpList.clearNames();
-                requestId++;
-                final String url = "http://andla.pythonanywhere.com/getnames/" + requestId + '/' + charSequence;
-                if (charSequence.length() != 0) {
-                    JsonObjectRequest jsonObjectRequest = new JsonObjectRequest
-                            (Request.Method.GET, url, null, new Response.Listener<JSONObject>() {
-
-                                @Override
-                                public void onResponse(JSONObject response) {
-                                    try {
-                                        if (response.getInt("id") == requestId) {
-                                            JSONArray results = response.getJSONArray("result");
-                                            for (int i = 0; i < results.length(); i++) {
-                                                names.add(results.get(i).toString());
-                                            }
-                                            popUpList.setNames(names);
-                                            interactiveSearcher.showPopupList();
-                                        }
-                                    } catch (JSONException e) {
-                                        e.printStackTrace();
-                                    }
-                                }
-                            }, new Response.ErrorListener() {
-
-                                @Override
-                                public void onErrorResponse(VolleyError error) {
-                                    Log.d("dee: ", error.toString());
-                                }
-                            });
-
-                    queue.add(jsonObjectRequest);
-                }
-            }
-
-            @Override
-            public void afterTextChanged(Editable editable) {
-                if (editable.toString().equals(" ")){
-                    //popUpList.clearNames();
-                }
-            }
-        });
     }
 }
